dependencies/requirements.txt
absl-py==1.2.0
addict==2.4.0
aiosignal==1.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.7
astunparse==1.6.3
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi==1.15.1
charset-normalizer==2.1.0
click==8.0.4
-e git+https://github.com/uzh-rpg/master_thesis_pietro_bonazzi.git@98e6a4b06666ef5c0581f35d6b60f424572eff8a#egg=clip&subdirectory=CLIP
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
ConfigArgParse==1.5.3
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.5
-e git+https://github.com/uzh-rpg/master_thesis_pietro_bonazzi.git@98e6a4b06666ef5c0581f35d6b60f424572eff8a#egg=DSS
easydict==1.9
entrypoints==0.4
executing==0.9.1
fastjsonschema==2.16.1
filelock==3.8.0
fire==0.4.0
flatbuffers==2.0.7
fonttools==4.34.4
frnn @ file:///home/bpietro/thesis/DSS/external/FRNN
frozenlist==1.3.1
ftfy==6.1.1
fvcore @ file:///home/conda/feedstock_root/build_artifacts/fvcore_1652361911092/work
gast==0.4.0
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.7.0
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
iniconfig==1.1.1
iopath==0.1.9
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.9
jsonschema==4.9.1
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-packaging==0.12.2
jupyter-server==1.18.1
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lpips==0.1.4
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.4
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.0
oauthlib==3.2.0
open3d==0.15.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.9.0
pluggy==1.0.0
plyfile==0.7.4
point-cloud-utils==0.25.1
portalocker @ file:///home/conda/feedstock_root/build_artifacts/portalocker_1657463947109/work
prefix-sum @ file:///home/bpietro/thesis/DSS/external/prefix_sum
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pymeshlab==2022.2.post2
pyparsing==3.0.9
pypoisson @ git+https://github.com/mmolero/pypoisson.git@605edfa8efbc1bfddf7fb7c8dc50385e412d2ad3
pyquaternion==0.9.9
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
pytorch-fid==0.2.1
pytorch3d==0.7.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1648757097602/work
pyzmq==23.2.0
ray==1.13.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==1.1.2
scipy==1.9.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tabulate @ file:///home/conda/feedstock_root/build_artifacts/tabulate_1655900837660/work
tenacity==8.0.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1657118200573/work
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.8.8
timm==0.6.7
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.3
torch==1.9.1
torch-batch-svd @ file:///home/bpietro/thesis/DSS/external/torch-batch-svd
torchvision==0.10.1
tornado==6.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1649051611147/work
traitlets==5.3.0
trimesh==3.6.34
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3==1.26.11
virtualenv==20.16.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.2
widgetsnbextension==3.6.1
wrapt==1.14.1
yacs @ file:///home/conda/feedstock_root/build_artifacts/yacs_1645705974477/work
zipp==3.8.1
